requirements.txt
Flask
Flask-SQLAlchemy
flask
requests
python-dotenv
gunicorn
Flask-Cors